+++ hahomematic/caches/visibility.py
@@ -9,7 +9,7 @@ import os
 from typing import Any, Final
 
 from hahomematic import central as hmcu, support as hms
-from hahomematic.const import CLICK_EVENTS, UN_IGNORE_WILDCARD, UTF8, Parameter, ParamsetKey
+from hahomematic.const import CLICK_EVENTS, ISO_8859_1, UN_IGNORE_WILDCARD, Parameter, ParamsetKey
 from hahomematic.model.custom import get_required_parameters
 from hahomematic.support import element_matches_key, reduce_args
 
@@ -706,7 +706,7 @@ class ParameterVisibilityCache:
                         self._storage_folder,
                         _FILE_CUSTOM_UN_IGNORE_PARAMETERS,
                     ),
-                    encoding=UTF8,
+                    encoding=ISO_8859_1,
                 ) as fptr:
                     for file_line in fptr.readlines():
                         if "#" not in file_line: