+++ hahomematic/support.py
@@ -32,10 +32,10 @@ from hahomematic.const import (
     HTMLTAG_PATTERN,
     IDENTIFIER_SEPARATOR,
     INIT_DATETIME,
+    ISO_8859_1,
     MAX_CACHE_AGE,
     NO_CACHE_ENTRY,
     PRIMARY_CLIENT_CANDIDATE_INTERFACES,
-    UTF8,
     CommandRxMode,
     ParamsetKey,
     RxMode,
@@ -75,7 +75,7 @@ def build_headers(
 ) -> list[tuple[str, str]]:
     """Build XML-RPC API header."""
     cred_bytes = f"{username}:{password}".encode()
-    base64_message = base64.b64encode(cred_bytes).decode(UTF8)
+    base64_message = base64.b64encode(cred_bytes).decode(ISO_8859_1)
     return [("Authorization", f"Basic {base64_message}")]